+++ src/index.html
@@ -22,10 +22,13 @@
             <ul id="lists" class="lists">
                     
             </ul>
-            <button class="clear-btn" type="button">Clear all completed</button>
+            <button id="clearCompletedBtn" type="button">Clear all completed</button>
+
         </section>
     
         <script type="module" src="index.js"></script>
+        <!-- <script type="module" src="./modules/crud.js"></script>
+        <script type="module" src="./modules/todos.js"></script> -->
     </body>
     </html>
