+++ src/index.js
@@ -1,38 +1,20 @@
-import AddList from './todos.js';
+import './styles.css';
+import AddList from './modules/todos.js';
+import CrudOperations from './modules/crud.js';
 
+const myCrud = new CrudOperations();
 const myList = new AddList();
-myList.displayList();
 
-function todoCheckbox() {
-  const checkboxes = document.querySelectorAll('.edit-text');
-  checkboxes.forEach((checkbox) => {
-    const index = checkbox.parentNode.querySelector('.editBtn').getAttribute('data-index');
-    const editInput = checkbox.parentNode.querySelector('.editBtn');
-    const { completed } = myList.todoDetails[index];
-
-    checkbox.checked = completed;
-    editInput.classList.toggle('completed', completed);
-
-    checkbox.addEventListener('change', (event) => {
-      const isChecked = event.target.checked;
-      editInput.classList.toggle('completed', isChecked);
-      myList.todoDetails[index].completed = isChecked;
-      localStorage.setItem('todoData', JSON.stringify(myList.todoDetails));
-    });
-  });
-}
+window.addEventListener('DOMContentLoaded', myList.displayList.bind(myList));
 
 const addButton = document.getElementById('add-button');
 addButton.addEventListener('click', () => {
   const task = document.getElementById('task').value.trim();
   const completed = false;
-  const index = myList.todoDetails.length + 1;
+  const index = myCrud.todoDetails.length + 1;
   if (task) {
-    myList.addRow(task, completed, index);
+    myCrud.addRow(task, completed, index);
     myList.displayList();
     document.getElementById('task').value = '';
-    todoCheckbox();
   }
 });
-
-window.addEventListener('DOMContentLoaded', todoCheckbox);
