+++ src/styles.css
@@ -71,7 +71,7 @@ button {
 }
 
 .clear-btn {
-  padding: 20px 0;
+  padding: 21px 0;
 }
 
 .clear-btn:hover {
